refactor: unify fixed/pod_dynamic codegen, delete dead tail-field code
Merge fixed.rs and pod_dynamic.rs into a single generate_account()
function that conditionally emits dynamic parts when any field has
PodDynField. Delete the old tail-field system remnants: read_tail_str,
read_tail_bytes, TailBytes, IdlTail, IdlType::Tail, and their codegen
match arms across all four client generators.
