+++ idl/src/types.rs
@@ -112,18 +112,12 @@ pub struct IdlDynVec {
 }
 
 #[derive(Serialize, Deserialize)]
-pub struct IdlTail {
-    pub element: String,
-}
-
-#[derive(Serialize, Deserialize)]
 #[serde(untagged)]
 pub enum IdlType {
     Primitive(String),
     Defined { defined: String },
     DynString { string: IdlDynString },
     DynVec { vec: IdlDynVec },
-    Tail { tail: IdlTail },
 }
 
 #[derive(Serialize, Deserialize)]
