+++ lang/src/instruction_data.rs
@@ -89,19 +89,6 @@ pub fn read_dynamic_vec<T, const PREFIX: usize>(
     Ok((slice, offset + byte_len))
 }
 
-/// Read a tail UTF-8 string (all remaining instruction data from `offset`).
-#[inline(always)]
-pub fn read_tail_str(data: &[u8], offset: usize) -> Result<&str, ProgramError> {
-    let bytes = &data[offset..];
-    core::str::from_utf8(bytes).map_err(|_| ProgramError::InvalidInstructionData)
-}
-
-/// Read tail bytes (all remaining instruction data from `offset`).
-#[inline(always)]
-pub fn read_tail_bytes(data: &[u8], offset: usize) -> &[u8] {
-    &data[offset..]
-}
-
 /// Read a length prefix from instruction data. The `PREFIX` const generic
 /// determines the byte width (1, 2, or 4). Monomorphized per variant so the
 /// match is eliminated at compile time.
